opentitantool: Fix issue with ISFB not treated as signed portion
We will need to upstream this change, but we need to land this now so we
can generate PAOs correctly.

The repack_with_paos.sh script calls `opentitantool image manifest
update --spx-key` which was error'ing with

```
Error: Invalid placement of signed extension 0x94ac01ec.
```

which is caused by OTT not recognizing that ISFB section is signed.
ISFB is supposed to be a signed section, which is clear from other
sections of the code.

BUG=none
TEST=./tools/repack_with_paos.sh succeeds now with ISFB in manifest.

Change-Id: Ida038540953e08c2ceef4eccbf5eb5c35ae1bb9a